requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-speech==2.28.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
langdetect==1.0.9
MarkupSafe==3.0.2
openai==1.57.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
